organise folders, scss changes
