+++ src/components/App/Footer/index.ts
