--- src/components/HappyCustomers/Customers.module.scss
@@ -1,4 +1,0 @@
-.imageBlock {
-  display: block;
-  width: 100%;
-}
