--- src/components/HappyCustomers/index.ts
@@ -1,1 +1,0 @@
-export { HappyCustomers } from "./HappyCustomers";
