+++ src/components/Home/Customers/Customers.module.scss
@@ -1,0 +1,9 @@
+.imageBlock {
+  display: block;
+  width: 100%;
+  border-radius: 0.25rem;
+}
+
+.widthBox {
+  width: 80%;
+}
