+++ src/components/Home/Customers/customerArray.ts
