+++ src/components/Home/PreviousJobs/PreviousJobs.module.scss
@@ -1,4 +1,5 @@
 .cardsContainer {
+  width: 80%;
   div figure {
     margin: 0;
     background-color: #343a40;
