+++ src/components/Home/index.ts
@@ -1,0 +1,3 @@
+export { Customers } from "./Customers/Customers";
+export { Introduction } from "./Introduction/Introduction";
+export { PreviousJobs } from "./PreviousJobs/PreviousJobs";
