--- src/components/Introduction/index.ts
@@ -1,1 +1,0 @@
-export { Introduction } from "./Introduction";
