--- src/components/PreviousJobs/index.ts
@@ -1,1 +1,0 @@
-export { PreviousJobs } from "./PreviousJobs";
