+++ src/pages/Home.tsx
@@ -1,14 +1,11 @@
-import { PicCarousel } from "components/PicCarousel";
-import { Introduction } from "components/Introduction";
-import { PreviousJobs } from "components/PreviousJobs";
-import { HappyCustomers } from "components/HappyCustomers";
+import { Customers, Introduction, PreviousJobs } from "components/Home";
 
 export default function Home() {
   return (
     <>
       <Introduction />
       <PreviousJobs />
-      <HappyCustomers />
+      <Customers />
     </>
   );
 }
