+++ src/styles/App.global.scss
@@ -2,3 +2,8 @@
   width: 90%;
   margin: 0 auto;
 }
+
+.headerText {
+  transform: rotate(-4deg);
+  padding-bottom: 3rem;
+}
